Better logging Rename the MessageBean to ExchangeMessageBean
