+++ service/src/main/java/eu/europa/ec/fisheries/uvms/exchange/service/bean/PluginServiceBean.java
@@ -40,7 +40,7 @@ import eu.europa.ec.fisheries.uvms.config.event.ConfigSettingUpdatedEvent;
 import eu.europa.ec.fisheries.uvms.config.exception.ConfigServiceException;
 import eu.europa.ec.fisheries.uvms.config.service.ParameterService;
 import eu.europa.ec.fisheries.uvms.config.service.UVMSConfigService;
-import eu.europa.ec.fisheries.uvms.exchange.message.consumer.ExchangeMessageConsumer;
+import eu.europa.ec.fisheries.uvms.exchange.message.consumer.ExchangeConsumer;
 import eu.europa.ec.fisheries.uvms.exchange.message.event.UpdatePluginSettingEvent;
 import eu.europa.ec.fisheries.uvms.exchange.message.event.carrier.ExchangeMessageEvent;
 import eu.europa.ec.fisheries.uvms.exchange.message.event.carrier.PluginMessageEvent;
@@ -80,7 +80,7 @@ public class PluginServiceBean implements PluginService {
     private ExchangeMessageProducer producer;
 
     @EJB
-    private  ExchangeMessageConsumer consumer;
+    private ExchangeConsumer consumer;
 
     @EJB
     private ParameterService parameterService;
